// src/interfaces/users.ts
export interface interfazUsuario {
    id: number;
    username: string;
    nombre: string;
    apellidos: string;
    email: string;
    dni: string;
    telefono: string;
    direccion: string;
    fecha_nacimiento: string;
    rol: string;
    visibilidad: boolean;
    fecha_actividad: string;
}